# src/accessi-module/openapi/accessi.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/accessi/email/reset-password-page/{token}": {
      "get": {
        "operationId": "serveResetPasswordPage",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token per il reset della password",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnUrl",
            "required": false,
            "in": "query",
            "description": "Url di ritorno della pagina",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pagina di reset password servita con successo"
          }
        },
        "summary": "Serve una pagina per il reset della password",
        "tags": [
          "Email"
        ]
      }
    },
    "/api/accessi/email/send-reset-password-email": {
      "post": {
        "operationId": "sendPasswordResetEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendResetPasswordEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "L'email di reset e stata gestita con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno durante l'invio dell'email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Invia una e-mail per il reset della password",
        "tags": [
          "Email"
        ]
      }
    },
...
    "/api/accessi/user/reset-password/{token}": {
      "get": {
        "operationId": "serveResetPasswordPageUser",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token per il reset della password",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Serve la pagina di reset password",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/get-users": {
      "get": {
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "Email dell'utente da cercare",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "codiceUtente",
            "required": false,
            "in": "query",
            "description": "Codice dell'utente da cercare",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeExtensionFields",
            "required": false,
            "in": "query",
            "description": "Includi extension fields",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "includeGrants",
            "required": false,
            "in": "query",
            "description": "Includi permessi",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Numero massimo di utenti restituiti (1-1000). Omesso = tutti.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Numero di utenti da saltare (paginazione).",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista utenti recuperata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenziali non valide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operazione non autorizzata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recupera la lista degli utenti",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/delete-user/{codiceUtente}": {
      "delete": {
        "description": "Imposta lo stato di un utente a eliminato senza rimuovere i record.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "description": "Codice identificativo dell'utente da eliminare",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utente eliminato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore nei parametri della richiesta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina un utente",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/force-password-reset/{codiceUtente}": {
      "post": {
        "description": "Riservato al superutente. Richiede un servizio email configurato; altrimenti risponde con ACCESSI_EMAIL_NOT_CONFIGURED.",
        "operationId": "forcePasswordReset",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "description": "Codice identificativo dell utente",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email di reset inviata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Servizio email non configurato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Forza l invio dell email di reset password a un utente",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/force-password-reset-legacy": {
      "post": {
        "description": "Riservato al superutente. Invia l email di reset a tutti gli utenti la cui password non e ancora nel formato moderno. Richiede un servizio email configurato.",
        "operationId": "forcePasswordResetLegacy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email di reset inviate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Servizio email non configurato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Forza il reset password per tutti gli utenti con password legacy",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/set-stato": {
      "post": {
        "operationId": "setStatoRegistrazione",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStatoRegistrazioneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stato registrazione aggiornato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore nei parametri della richiesta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Imposta lo stato di registrazione di un utente",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/register": {
      "post": {
        "description": "Disponibile soltanto quando publicRegistration.enabled e true. Per utenti SSO usare il provisioning master nel tag Federated authentication.",
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Dati necessari per la registrazione dell'utente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utente registrato con successo. Restituisce il codice utente e invia una mail di reset password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore nella registrazione.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server durante la registrazione o l'invio email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Registra un nuovo utente",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/create-managed-user": {
      "post": {
        "description": "Riservato a superutente. Crea un utente in attesa di impostazione password e invia la mail di reset.",
        "operationId": "createManagedUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea un utente locale da amministrazione",
        "tags": [
          "User"
        ]
      }
    },
    "/api/accessi/user/update-user/{codiceUtente}": {
      "put": {
        "operationId": "updateUtente",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "description": "Codice identificativo dell'utente da aggiornare",
            "schema": {
              "example": 123,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dati aggiornati dell'utente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utente aggiornato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore nell'aggiornamento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operazione non autorizzata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aggiorna un utente esistente",
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Accessi API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SendResetPasswordEmailRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "L'email dell'utente che richiede il reset.",
            "example": "mario.rossi@dev.it"
          },
          "htmlMail": {
            "type": "string",
            "description": "HTML personalizzato della mail di reset, se gestito dal chiamante.",
            "example": "<html><body>Reset password</body></html>"
          }
        },
        "required": [
          "email"
        ]
      },
      "ActionResponse": {
        "type": "object",
        "properties": {
          "severity": {
            "type": "string",
            "example": "success"
          },
          "status": {
            "type": "number",
            "example": 200
          },
          "statusCode": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Operazione completata con successo."
          }
        },
        "required": [
          "severity",
          "status",
          "statusCode",
          "message"
        ]
      },
...
      "GetUsersResponse": {
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/Status"
          },
          "Message": {
            "type": "string",
            "example": "Dati recuperati con successo.",
            "description": "Messaggio informativo restituito dalla libreria quando presente."
          },
          "Result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetUsersResult"
            }
          }
        },
        "required": [
          "Status",
          "Result"
        ]
      },
...
    }
  }
}